requirements.txt
# oTree-may-overwrite-this-file
# IF YOU MODIFY THIS FILE, remove this comment.
# otherwise, oTree will automatically overwrite it.
otree>=5.0.0a21
psycopg2>=2.8.4
arm-mango==1.4.3
ipython==8.18.1